# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ERP Intelligence Dashboard API",
    "description": "Backend API untuk ERP Intelligence Dashboard ISP. Menyajikan wawasan analitis, model prediksi, dan data terstruktur Feature Store.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/ml/models/churn/train": {
      "post": {
        "tags": [
          "Machine Learning & Predictive Intelligence"
        ],
        "summary": "Train or Retrain Customer Churn Model (XGBoost)",
        "description": "Memicu siklus pelatihan model Customer Churn berbasis XGBoost. Mengambil data dari feature_store.feature_customer_churn, menerapkan anti-leakage temporal split, menghitung metrik evaluasi (ROC-AUC, F1, PR-AUC), menyimpan artifact terversi ke local registry, dan memperbarui pointer active_model.json.",
        "operationId": "train_churn_model_endpoint_api_v1_ml_models_churn_train_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ModelTrainRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope_ModelTrainResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ml/predictions/churn/batch": {
      "post": {
        "tags": [
          "Machine Learning & Predictive Intelligence"
        ],
        "summary": "Execute Batch Churn Prediction for a Snapshot",
        "description": "Menjalankan inferensi batch untuk semua pelanggan pada snapshot tertentu. Menghasilkan probabilitas, risk level (LOW/MED/HIGH), risk rank, dan top 3 driver risiko. Hasil disimpan secara idempotent (ON CONFLICT DO UPDATE) di feature_store.prediction_customer_churn.",
        "operationId": "predict_churn_batch_endpoint_api_v1_ml_predictions_churn_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/BatchPredictionRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope_BatchPredictionResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ml/predictions/churn/high-risk": {
      "get": {
        "tags": [
          "Machine Learning & Predictive Intelligence"
        ],
        "summary": "Get Top High-Risk Customers for Retention Focus",
        "description": "Mengembalikan daftar pelanggan berisiko tinggi terurut berdasarkan probabilitas churn tertinggi. Menyertakan nama pelanggan, kota, status, serta Top 3 faktor pemicu risiko (SHAP drivers).",
        "operationId": "get_high_risk_customers_endpoint_api_v1_ml_predictions_churn_high_risk_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maksimum jumlah pelanggan yang ditampilkan",
              "default": 20,
              "title": "Limit"
            },
            "description": "Maksimum jumlah pelanggan yang ditampilkan"
          },
          {
            "name": "min_probability",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 1.0,
              "minimum": 0.0,
              "description": "Ambang batas probabilitas minimal",
              "default": 0.7,
              "title": "Min Probability"
            },
            "description": "Ambang batas probabilitas minimal"
          },
          {
            "name": "snapshot_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Tanggal snapshot (opsional, default: terbaru)",
              "title": "Snapshot Date"
            },
            "description": "Tanggal snapshot (opsional, default: terbaru)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope_List_CustomerRiskProfile__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ml/predictions/churn/customer/{customer_id}": {
      "get": {
        "tags": [
          "Machine Learning & Predictive Intelligence"
        ],
        "summary": "Get Individual Customer Churn Risk Profile",
        "description": "Mengambil profil risiko churn dan detail driver pemicu untuk satu pelanggan tertentu.",
        "operationId": "get_customer_prediction_endpoint_api_v1_ml_predictions_churn_customer__customer_id__get",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Customer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope_CustomerRiskProfile_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dashboard/overview": {
      "get": {
        "tags": [
          "Insights & Executive Dashboard"
        ],
        "summary": "Get Executive Overview Dashboard Insights",
        "description": "Mengambil wawasan ringkasan eksekutif tingkat tinggi lintas modul bisnis ISP.",
        "operationId": "get_dashboard_overview_api_v1_dashboard_overview_get",
        "parameters": [
          {
            "name": "as_of_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Tanggal snapshot (YYYY-MM-DD)",
              "title": "As Of Date"
            },
            "description": "Tanggal snapshot (YYYY-MM-DD)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope_InsightPackage_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/insights/{module}": {
      "get": {
        "tags": [
          "Insights & Executive Dashboard"
        ],
        "summary": "Get Domain Module Insights Package",
        "description": "Mengambil paket wawasan terstruktur untuk modul bisnis tertentu. Mencakup ringkasan eksekutif kritis LLM, kartu KPI, narasi temuan berbasis alasan agen, dan spesifikasi grafik Vega-Lite v5 yang terikat pada data aktual.",
        "operationId": "get_module_insight_api_v1_insights__module__get",
        "parameters": [
          {
            "name": "module",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Module"
            }
          },
          {
            "name": "as_of_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Tanggal snapshot (YYYY-MM-DD)",
              "title": "As Of Date"
            },
            "description": "Tanggal snapshot (YYYY-MM-DD)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope_InsightPackage_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchPredictionRequest": {
        "properties": {
          "snapshot_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot Date",
            "description": "Tanggal snapshot fitur (YYYY-MM-DD). Jika null, otomatis gunakan snapshot terbaru."
          },
          "force_refresh": {
            "type": "boolean",
            "title": "Force Refresh",
            "description": "Jika true, lakukan overwrite idempotent pada prediksi yang sudah ada",
            "default": true
          }
        },
        "type": "object",
        "title": "BatchPredictionRequest"
      },
...
      "ErrorDetail": {
        "properties": {
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field",
            "description": "Nama field yang mengalami kegagalan validasi"
          },
          "issue": {
            "type": "string",
            "title": "Issue",
            "description": "Deskripsi permasalahan validasi"
          }
        },
        "type": "object",
        "required": [
          "issue"
        ],
        "title": "ErrorDetail"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelTrainRequest": {
        "properties": {
          "model_version": {
            "type": "string",
            "title": "Model Version",
            "description": "Versi model semantic yang akan dilatih",
            "default": "1.0.0"
          },
          "test_size_ratio": {
            "type": "number",
            "maximum": 0.5,
            "minimum": 0.05,
            "title": "Test Size Ratio",
            "description": "Proporsi data temporal untuk validasi OOT",
            "default": 0.2
          },
          "hyperparameters": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hyperparameters",
            "description": "Custom hyperparameter dictionary untuk XGBoost"
          },
          "set_as_active": {
            "type": "boolean",
            "title": "Set As Active",
            "description": "Jika true, langsung jadikan model ini sebagai active inference model",
            "default": true
          }
        },
        "type": "object",
        "title": "ModelTrainRequest"
      },
      "ModelTrainResponse": {
        "properties": {
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "trained_at": {
            "type": "string",
            "title": "Trained At"
          },
          "total_samples": {
            "type": "integer",
            "title": "Total Samples"
          },
          "metrics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metrics"
          },
          "top_global_features": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Top Global Features"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "model_name",
          "model_version",
          "status",
          "trained_at",
          "total_samples",
          "metrics",
          "top_global_features",
          "is_active"
        ],
        "title": "ModelTrainResponse"
      },
...
      "ResponseEnvelope_BatchPredictionResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Status keberhasilan operasi"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BatchPredictionResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Payload data utama respons"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta",
            "description": "Metadata pendukung"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResponseError"
              },
              {
                "type": "null"
              }
            ],
            "description": "Objek error jika operasi gagal"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ResponseEnvelope[BatchPredictionResponse]"
      },
      "ResponseEnvelope_CustomerRiskProfile_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Status keberhasilan operasi"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CustomerRiskProfile"
              },
              {
                "type": "null"
              }
            ],
            "description": "Payload data utama respons"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta",
            "description": "Metadata pendukung"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResponseError"
              },
              {
                "type": "null"
              }
            ],
            "description": "Objek error jika operasi gagal"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ResponseEnvelope[CustomerRiskProfile]"
      },
...
      "ResponseEnvelope_InsightPackage_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Status keberhasilan operasi"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InsightPackage"
              },
              {
                "type": "null"
              }
            ],
            "description": "Payload data utama respons"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta",
            "description": "Metadata pendukung"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResponseError"
              },
              {
                "type": "null"
              }
            ],
            "description": "Objek error jika operasi gagal"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ResponseEnvelope[InsightPackage]"
      },
...
      "ResponseEnvelope_List_CustomerRiskProfile__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Status keberhasilan operasi"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/CustomerRiskProfile"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "Payload data utama respons"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta",
            "description": "Metadata pendukung"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResponseError"
              },
              {
                "type": "null"
              }
            ],
            "description": "Objek error jika operasi gagal"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ResponseEnvelope[List[CustomerRiskProfile]]"
      },
...
      "ResponseEnvelope_ModelTrainResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Status keberhasilan operasi"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelTrainResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Payload data utama respons"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta",
            "description": "Metadata pendukung"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResponseError"
              },
              {
                "type": "null"
              }
            ],
            "description": "Objek error jika operasi gagal"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ResponseEnvelope[ModelTrainResponse]"
      },
...
      "ResponseError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Katalog kode error standar (e.g. VALIDATION_ERROR, NOT_FOUND)"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Pesan deskriptif error yang ramah pengguna"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details",
            "description": "Rincian error per-field"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ResponseError"
      },
      "ResponseMeta": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "Pengenal unik transaksi request (UUIDv4) untuk tracing"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "Waktu pembuatan respons dalam format ISO-8601 UTC"
          },
          "version": {
            "type": "string",
            "title": "Version",
            "description": "Versi API yang melayani request",
            "default": "v1"
          },
          "pagination": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaginationMeta"
              },
              {
                "type": "null"
              }
            ],
            "description": "Metadata pagination (jika respons mengembalikan daftar data)"
          }
        },
        "type": "object",
        "title": "ResponseMeta"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}